+++ src/components/buttons/sucess/index.js
@@ -1,4 +1,4 @@
-import Button from '../../button'
+import Button from '../button'
 import './sucess.css'
 
 const SucessButton = ({ icon, onClick, text, type }) => {
