+++ src/components/togglerow/index.js
@@ -2,7 +2,7 @@ import { faAngleDown, faPen } from '@fortawesome/free-solid-svg-icons'
 import { useState } from 'react'
 import './style.css'
 import ActionButton from '../buttons/action'
-import Button from '../button'
+import Button from '../buttons/button'
 import { Status } from '../status'
 import { client } from '../../service'
 import { useNavigate } from 'react-router-dom'
